add role list unit test
