+++ cmd/karavictl/cmd/role_get_test.go
@@ -127,7 +127,6 @@ func TestRoleGetHandler(t *testing.T) {
 					if err != nil {
 						t.Fatal(err)
 					}
-					t.Log(resp)
 					return nil
 				},
 			}, nil
